Restructure baseline of retriever and document_store
