+++ modules/ml/__init__.py
@@ -1,19 +1,19 @@
-import logging
+# import logging
 
-import pandas as pd
-from modules.ml.schema import Document, Label, MultiLabel
-from modules.ml.finder import Finder
-from modules.ml.pipeline import Pipeline
+# import pandas as pd
+# from modules.ml.schema import Document, Label, MultiLabel
+# from modules.ml.finder import Finder
+# from modules.ml.pipeline import Pipeline
 
-pd.options.display.max_colwidth = 80
+# pd.options.display.max_colwidth = 80
 
-logger = logging.getLogger(__name__)
+# logger = logging.getLogger(__name__)
 
-logging.getLogger('farm').setLevel(logging.WARNING)
-logging.getLogger('farm.utils').setLevel(logging.INFO)
-logging.getLogger('farm.infer').setLevel(logging.INFO)
-logging.getLogger('transformers').setLevel(logging.WARNING)
-logging.getLogger('farm.eval').setLevel(logging.INFO)
-logging.getLogger('farm.modeling.optimization').setLevel(logging.INFO)
+# logging.getLogger('farm').setLevel(logging.WARNING)
+# logging.getLogger('farm.utils').setLevel(logging.INFO)
+# logging.getLogger('farm.infer').setLevel(logging.INFO)
+# logging.getLogger('transformers').setLevel(logging.WARNING)
+# logging.getLogger('farm.eval').setLevel(logging.INFO)
+# logging.getLogger('farm.modeling.optimization').setLevel(logging.INFO)
