+++ modules/ml/preprocessor/utils.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from typing import Callable, Dict, List, Optional, Tuple, Union, Generator
 import json
 
-from farm.data_handler.utils import http_get
+# from farm.data_handler.utils import http_get
 
 from modules.ml.file_converter.base import BaseConverter
 from modules.ml.file_converter.docx import DocxToTextConverter
